refactor: remove filter.go
